remove scoreview file
