+++ src/index.js
@@ -1,6 +1,7 @@
 import './style.css';
-import viewScoreBoard from './modules/scoreView.js';
+// import viewScoreBoard from './modules/scoreView.js';
 
+const scoreContainer = document.querySelector('.player-score-container');
 const playerScores = [
   { name: 'name', score: 100 },
   { name: 'name', score: 20 },
@@ -11,4 +12,17 @@ const playerScores = [
   { name: 'name', score: 42 },
 ];
 
-viewScoreBoard(playerScores);
+// dynamically populate todo player name and score
+const viewScoreBoard = (playerScores) => {
+  let score = '';
+  playerScores.forEach((player) => {
+    score += `
+      <li class="each-player">
+        ${player.name}: ${player.score}
+      </li>
+    `;
+  });
+  scoreContainer.innerHTML = score;
+};
+
+viewScoreBoard();
