+++ src/modules/scoreView.js
@@ -1,16 +1,16 @@
-const scoreContainer = document.querySelector('.player-score-container');
+// const scoreContainer = document.querySelector('.player-score-container');
 
-// dynamically populate todo player name and score
-const viewScoreBoard = (playerScores) => {
-  let score = '';
-  playerScores.forEach((player) => {
-    score += `
-      <li class="each-player">
-        ${player.name}: ${player.score}
-      </li>
-    `;
-  });
-  scoreContainer.innerHTML = score;
-};
+// // dynamically populate todo player name and score
+// const viewScoreBoard = (playerScores) => {
+//   let score = '';
+//   playerScores.forEach((player) => {
+//     score += `
+//       <li class="each-player">
+//         ${player.name}: ${player.score}
+//       </li>
+//     `;
+//   });
+//   scoreContainer.innerHTML = score;
+// };
 
-export default viewScoreBoard;
+// export default viewScoreBoard;
